services/vanna/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
groq>=0.6.0
pydantic==2.9.0
sqlalchemy==2.0.23